+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [1.3.0] - 2026-02-18
+
+### Added
+- Added correlate only scripting and fieldMapping
+- Added permission for drives (permission, grant, revoke and subPermissions)
+- Added resources for drives
+
 ## [1.2.0] - 2025-02-09
 
 ### Added